@@ -16,17 +16,17 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "axios": "^0.26.1",
-    "commander": "^9.1.0",
+    "axios": "^0.27.2",
+    "commander": "^9.3.0",
     "console-grid": "^1.0.17",
     "crypto-js": "^4.1.1",
     "date-fns": "^2.28.0",
-    "dotenv": "^16.0.0",
+    "dotenv": "^16.0.1",
     "download": "^8.0.0",
     "form-data": "^4.0.0",
     "got": "^11.8.3",
-    "http-server": "^14.1.0",
-    "jsdom": "^19.0.0",
+    "http-server": "^14.1.1",
+    "jsdom": "^20.0.0",
     "md5": "^2.3.0",
     "png-js": "^1.0.0",
     "prettytable": "^0.3.1",
@@ -34,13 +34,17 @@
     "request": "^2.88.2",
     "tough-cookie": "^4.0.0",
     "ts-md5": "^1.2.11",
-    "tslib": "^2.3.1",
+    "tslib": "^2.4.0",
     "tunnel": "0.0.6",
-    "ws": "^8.5.0"
+    "ws": "^8.8.0"
   },
   "devDependencies": {
-    "@types/node": "^17.0.23",
-    "ts-node": "^10.7.0",
-    "typescript": "^4.6.3"
+    "@types/express": "^4.17.13",
+    "@types/node": "^18.0.0",
+    "express": "^4.18.1",
+    "gulp": "^4.0.2",
+    "gulp-typescript": "^6.0.0-alpha.1",
+    "ts-node": "^10.8.2",
+    "typescript": "^4.7.4"
   }
 }
