@@ -1,10 +1,15 @@
-/*
-此文件为Node.js专用。其他用户请忽略
- */
-//此处填写京东账号cookie。
-let CookieJDs = [
-]
-// 判断环境变量里面是否有京东ck
+const fs = require('fs')
+const dotenv = require('dotenv')
+
+let CookieJDs = []
+
+try {
+  process.chdir(__dirname)
+  fs.accessSync('.env', fs.constants.R_OK)
+  dotenv.config()
+} catch (e) {
+}
+
 if (process.env.JD_COOKIE) {
   if (process.env.JD_COOKIE.indexOf('&') > -1) {
     CookieJDs = process.env.JD_COOKIE.split('&');
@@ -14,19 +19,15 @@ if (process.env.JD_COOKIE) {
     CookieJDs = [process.env.JD_COOKIE];
   }
 }
-if (JSON.stringify(process.env).indexOf('GITHUB')>-1) {
-  console.log(`请勿使用github action运行此脚本,无论你是从你自己的私库还是其他哪里拉取的源代码，都会导致我被封号\n`);
-  !(async () => {
-    await require('./sendNotify').sendNotify('提醒', `请勿使用github action、滥用github资源会封我仓库以及账号`)
-    await process.exit(0);
-  })()
+if (JSON.stringify(process.env).indexOf('GITHUB') > -1) {
+  process.exit(0);
 }
 CookieJDs = [...new Set(CookieJDs.filter(item => !!item))]
-console.log(`\n====================共${CookieJDs.length}个京东账号Cookie=========\n`);
-console.log(`==================脚本执行- 北京时间(UTC+8)：${new Date(new Date().getTime() + new Date().getTimezoneOffset()*60*1000 + 8*60*60*1000).toLocaleString()}=====================\n`)
-if (process.env.JD_DEBUG && process.env.JD_DEBUG === 'false') console.log = () => {};
+if (!require.main.filename.includes('.ts')) {
+  console.log(`\n====================共${CookieJDs.length}个京东账号Cookie=========\n`);
+  console.log(`================== ${new Date().toLocaleString('zh-CN', {timeZone: 'Asia/Shanghai'})} =====================\n`)
+}
 for (let i = 0; i < CookieJDs.length; i++) {
-  if (!CookieJDs[i].match(/pt_pin=(.+?);/) || !CookieJDs[i].match(/pt_key=(.+?);/)) console.log(`\n提示:京东cookie 【${CookieJDs[i]}】填写不规范,可能会影响部分脚本正常使用。正确格式为: pt_key=xxx;pt_pin=xxx;（分号;不可少）\n`);
   const index = (i + 1 === 1) ? '' : (i + 1);
   exports['CookieJD' + index] = CookieJDs[i].trim();
 }
