@@ -1,21 +1,11 @@
-/*
-京东种豆得豆互助码
-此文件为Node.js专用。其他用户请忽略
-支持京东N个账号
+/**
+ * 京东种豆得豆互助码
+ * 同一个京东账号的好友互助码用@符号隔开,不同京东账号之间用&符号或者换行隔开,下面给一个示例
+ * 如: 京东账号1的shareCode1@京东账号1的shareCode2&京东账号2的shareCode1@京东账号2的shareCode2
  */
-//云服务器腾讯云函数等NOde.js用户在此处填写东东萌宠的好友码。
-// 同一个京东账号的好友互助码用@符号隔开,不同京东账号之间用&符号或者换行隔开,下面给一个示例
-// 如: 京东账号1的shareCode1@京东账号1的shareCode2&京东账号2的shareCode1@京东账号2的shareCode2
-let PlantBeanShareCodes = [
-]
 
-// 从日志获取互助码
-// const logShareCodes = require('./utils/jdShareCodes');
-// if (logShareCodes.PLANT_BEAN_SHARECODES.length > 0 && !process.env.PLANT_BEAN_SHARECODES) {
-//   process.env.PLANT_BEAN_SHARECODES = logShareCodes.PLANT_BEAN_SHARECODES.join('&');
-// }
+let PlantBeanShareCodes = []
 
-// 判断github action里面是否有种豆得豆互助码
 if (process.env.PLANT_BEAN_SHARECODES) {
   if (process.env.PLANT_BEAN_SHARECODES.indexOf('&') > -1) {
     console.log(`您的种豆互助码选择的是用&隔开\n`)
@@ -27,7 +17,7 @@ if (process.env.PLANT_BEAN_SHARECODES) {
     PlantBeanShareCodes = process.env.PLANT_BEAN_SHARECODES.split();
   }
 } else {
-  console.log(`由于您环境变量(PLANT_BEAN_SHARECODES)里面未提供助力码，故此处运行将会给脚本内置的码进行助力，请知晓！`)
+  console.log('PLANT_BEAN_SHARECODES undefined')
 }
 for (let i = 0; i < PlantBeanShareCodes.length; i++) {
   const index = (i + 1 === 1) ? '' : (i + 1);
