@@ -5,13 +5,12 @@
  */
 
 import axios from 'axios'
-import USER_AGENT, {get, randomNumString, requireConfig, wait} from './TS_USER_AGENTS'
+import USER_AGENT, {get, randomNumString, getCookie, wait} from './TS_USER_AGENTS'
 
 let cookie: string = '', res: any = '', UserName: string, index: number, uuid: string
-// let shareCodeSelf: { shareCode: string, groupCode: string, activeId: string }[] = [], shareCode: { shareCode: string, groupCode: string, activeId: string }[] = [], shareCodeHW: { shareCode: string, groupCode: string, activeId: string }[] = []
 
 !(async () => {
-  let cookiesArr: string[] = await requireConfig()
+  let cookiesArr: string[] = await getCookie()
   for (let i = 0; i < cookiesArr.length; i++) {
     cookie = cookiesArr[i]
     UserName = decodeURIComponent(cookie.match(/pt_pin=([^;]*)/)![1])
@@ -25,12 +24,12 @@ let cookie: string = '', res: any = '', UserName: string, index: number, uuid: s
       'Referer': 'https://h5.m.jd.com/',
       'Cookie': cookie
     }
-    res = await get('https://api.m.jd.com/client.action?functionId=initForTurntableFarm&body=%7B%22version%22%3A4%2C%22channel%22%3A1%7D&appid=wh5', '', headers)
+    res = await get('https://api.m.jd.com/client.action?functionId=initForTurntableFarm&body=%7B%22version%22%3A4%2C%22channel%22%3A1%7D&appid=wh5', headers)
     let times: number = res.remainLotteryTimes
     console.log('剩余抽奖机会:', times)
     for (let j = 0; j < times; j++) {
       console.log('开始抽奖...')
-      res = await get('https://api.m.jd.com/client.action?functionId=lotteryForTurntableFarm&body=%7B%22type%22%3A1%2C%22version%22%3A4%2C%22channel%22%3A1%7D&appid=wh5', '', headers)
+      res = await get('https://api.m.jd.com/client.action?functionId=lotteryForTurntableFarm&body=%7B%22type%22%3A1%2C%22version%22%3A4%2C%22channel%22%3A1%7D&appid=wh5', headers)
       if (res.code === '0') {
         if (res.type === 'thanks') {
           console.log('抽奖成功，获得：狗屁')
@@ -74,41 +73,6 @@ let cookie: string = '', res: any = '', UserName: string, index: number, uuid: s
       }
       await wait(2000)
     }
-
-    // 抢京豆
-    /*
-    res = await qjd('signBeanGroupStageIndex', {"monitor_refer": "", "rnVersion": "3.9", "fp": "-1", "shshshfp": "-1", "shshshfpa": "-1", "referUrl": "-1", "userAgent": "-1", "jda": "-1", "monitor_source": "bean_m_bean_index"})
-    if (!res.data.shareCode) {
-      console.log('抢京豆 init...')
-      res = await qjd('signGroupHit', {"activeType": 2})
-      await wait(1000)
-      res = await qjd('signBeanGroupStageIndex', {"monitor_refer": "", "rnVersion": "3.9", "fp": "-1", "shshshfp": "-1", "shshshfpa": "-1", "referUrl": "-1", "userAgent": "-1", "jda": "-1", "monitor_source": "bean_m_bean_index"})
-    }
-    if (res.data.shareCode) {
-      console.log('助力码', res.data.shareCode)
-      shareCodeSelf.push({shareCode: res.data.shareCode, groupCode: res.data.groupCode, activeId: res.data.jklInfo.keyId})
-      await wait(2000)
-    }
-  }
-
-  console.log('内部助力', shareCodeSelf)
-  for (let i = 0; i < cookiesArr.length; i++) {
-    if (shareCodeHW.length === 0) {
-      await getShareCodeHW()
-    }
-    shareCode = [...shareCodeSelf, ...shareCodeHW]
-    uuid = randomString(40)
-    cookie = cookiesArr[i]
-    UserName = decodeURIComponent(cookie.match(/pt_pin=([^;]*)/)![1])
-    for (let code of shareCode) {
-      console.log(`${UserName} 去助力 ${code.shareCode}`)
-      res = await qjd('signGroupHelp', {"activeType": 2, "groupCode": code.groupCode, "shareCode": code.shareCode, "activeId": code.activeId, "source": "guest"})
-      console.log(res.data?.helpToast)
-      if (res.data.respCode === 'SG209')
-        break
-      await wait(2000)
-    }
-  */
   }
 })()
 
@@ -125,15 +89,3 @@ async function api(fn: string, body: object) {
     })
   return data
 }
-
-/*async function qjd(fn: string, body?: object) {
-  let {data} = await axios.get(`https://api.m.jd.com/client.action?functionId=${fn}&body=${encodeURIComponent(JSON.stringify(body))}&appid=ld&client=apple&clientVersion=10.0.8&uuid=${uuid}&openudid=${uuid}`, {
-    headers: {
-      'Host': 'api.m.jd.com',
-      'User-Agent': USER_AGENT,
-      'Referer': 'https://h5.m.jd.com/rn/3MQXMdRUTeat9xqBSZDSCCAE9Eqz/index.html',
-      'Cookie': cookie
-    }
-  })
-  return data
-}*/
