@@ -30,7 +30,7 @@ console.log('\n====================Hello World====================\n')
 //Node.js用户请在jdCookie.js处填写京东ck;
 //ios等软件用户直接用NobyDa的jd cookie
 let jdNotify = true;  //是否开启静默运行。默认true开启
-let cookiesArr = [], cookie = '', jdPlantBeanShareArr = [], isBox = false, notify, newShareCodes, option, message,
+let cookiesArr = [], cookie = '', notify, newShareCodes, option, message,
   subTitle;
 //京东接口地址
 const JD_API_HOST = 'https://api.m.jd.com/client.action';
@@ -106,14 +106,13 @@ async function jdPlantBean() {
       currentRoundId = roundList[1].roundId;//本期的roundId
       lastRoundId = roundList[0].roundId;//上期的roundId
       awardState = roundList[0].awardState;
-      $.taskList = $.plantBeanIndexResult.data.taskList;
+      $.taskList = $.plantBeanIndexResult.data?.taskList;
       subTitle = `【京东昵称】${$.plantBeanIndexResult.data.plantUserInfo.plantNickName}`;
       message += `【上期时间】${roundList[0].dateDesc.replace('上期 ', '')}\n`;
       message += `【上期成长值】${roundList[0].growth}\n`;
       await receiveNutrients();//定时领取营养液
       await doHelp();//助力
       await doTask();//做日常任务
-      await doEgg();
       await stealFriendWater();
       await doCultureBean();
       await doGetReward();
@@ -133,7 +132,7 @@ async function jdPlantBean() {
 function runTimes() {
   return new Promise((resolve, reject) => {
     $.get({
-      url: `https://api.jdsharecode.xyz/api/runTimes0407?activityId=bean&sharecode=${$.myPlantUuid}`
+      url: `https://api.jdsharecode.xyz/api/runTimes0701?activityId=bean&sharecode=${$.myPlantUuid}`
     }, (err, resp, data) => {
       if (err) {
         console.log('上报失败', err)
@@ -230,25 +229,6 @@ async function stealFriendWater() {
     }
   } else {
     console.log(`$.stealFriendList 异常： ${JSON.stringify($.stealFriendList)}`)
-  }
-}
-
-async function doEgg() {
-  await egg();
-  if ($.plantEggLotteryRes && $.plantEggLotteryRes.code === '0') {
-    if ($.plantEggLotteryRes.data.restLotteryNum > 0) {
-      const eggL = new Array($.plantEggLotteryRes.data.restLotteryNum).fill('');
-      console.log(`目前共有${eggL.length}次扭蛋的机会`)
-      for (let i = 0; i < eggL.length; i++) {
-        console.log(`开始第${i + 1}次扭蛋`);
-        await plantEggDoLottery();
-        console.log(`天天扭蛋成功：${JSON.stringify($.plantEggDoLotteryResult)}`);
-      }
-    } else {
-      console.log('暂无扭蛋机会')
-    }
-  } else {
-    console.log('查询天天扭蛋的机会失败' + JSON.stringify($.plantEggLotteryRes))
   }
 }
 
@@ -410,7 +390,7 @@ async function doTask() {
 function showTaskProcess() {
   return new Promise(async resolve => {
     await plantBeanIndex();
-    $.taskList = $.plantBeanIndexResult.data.taskList;
+    $.taskList = $.plantBeanIndexResult.data?.taskList;
     if ($.taskList && $.taskList.length > 0) {
       console.log("     任务   进度");
       for (let item of $.taskList) {
@@ -432,8 +412,7 @@ async function doHelp() {
     }
     await helpShare(plantUuid);
     if ($.helpResult && $.helpResult.code === '0') {
-      // console.log(`助力好友结果: ${JSON.stringify($.helpResult.data.helpShareRes)}`);
-      if ($.helpResult.data.helpShareRes) {
+      if ($.helpResult.data?.helpShareRes) {
         if ($.helpResult.data.helpShareRes.state === '1') {
           console.log(`助力好友${plantUuid}成功`)
           console.log(`${$.helpResult.data.helpShareRes.promptText}\n`);
@@ -633,7 +612,15 @@ function requireConfig() {
     notify = $.isNode() ? require('./sendNotify') : '';
     //Node.js用户请在jdCookie.js处填写京东ck;
     const jdCookieNode = $.isNode() ? require('./jdCookie.js') : '';
-    const jdPlantBeanShareCodes = $.isNode() ? require('./jdPlantBeanShareCodes.js') : '';
+    let jdPlantBeanShareCodes = ''
+    if ($.isNode()) {
+      try {
+        const dotenv = require('dotenv');
+        dotenv.config();
+        jdPlantBeanShareCodes = require('./jdPlantBeanShareCodes')
+      } catch (e) {
+      }
+    }
     //IOS等用户直接用NobyDa的jd cookie
     if ($.isNode()) {
       Object.keys(jdCookieNode).forEach((item) => {
