@@ -1,61 +1,69 @@
-import axios from 'axios'
-import {requireConfig, wait, o2s} from './TS_USER_AGENTS'
-import {requestAlgo, geth5st} from "./utils/V3";
-import {SHA256} from "crypto-js";
+/**
+ * 极速版-签到+提现
+ * cron: 45 0 * * *
+ */
 
-let cookie: string = '', res: any = '', UserName: string = ''
+import {H5ST} from "./utils/h5st"
+import {JDHelloWorld, User} from "./TS_JDHelloWorld";
 
-!(async () => {
-  let cookiesArr: any = await requireConfig()
-  for (let [index, value] of cookiesArr.entries()) {
+class Speed_Sign extends JDHelloWorld {
+  cookie: string
+  h5stTool: H5ST
+
+  constructor() {
+    super();
+  }
+
+  async init() {
+    await this.run(new Speed_Sign())
+  }
+
+
+  async api(fn: string, body: object) {
+    let timestamp: number = Date.now()
+    let h5st: string = this.h5stTool.__genH5st({
+      appid: 'activities_platform',
+      body: JSON.stringify(body),
+      client: 'H5',
+      clientVersion: '1.0.0',
+      functionId: fn,
+      t: timestamp.toString()
+    })
+    return await this.post('https://api.m.jd.com/', `functionId=${fn}&body=${encodeURIComponent(JSON.stringify(body))}&t=${timestamp}&appid=activities_platform&client=H5&clientVersion=1.0.0&h5st=${h5st}`, {
+      'Host': 'api.m.jd.com',
+      'User-Agent': 'jdltapp;android;3.8.16;',
+      'Origin': 'https://daily-redpacket.jd.com',
+      'Referer': 'https://daily-redpacket.jd.com/',
+      'Content-Type': 'application/x-www-form-urlencoded',
+      'Cookie': this.cookie
+    })
+  }
+
+  async main(user: User) {
+    this.cookie = user.cookie
+    this.h5stTool = new H5ST("15097", "jdltapp;", process.env.FP_15097 ?? "");
+    await this.h5stTool.__genAlgo()
+    let res: any = await this.api('apSignIn_day', {"linkId": "9WA12jYGulArzWS7vcrwhw", "serviceName": "dayDaySignGetRedEnvelopeSignService", "business": 1})
     try {
-      await requestAlgo('15097', 'jdltapp;')
-      cookie = value
-      UserName = decodeURIComponent(cookie.match(/pt_pin=([^;]*)/)![1])
-      console.log(`\n开始【京东账号${index + 1}】${UserName}\n`)
-
-      res = await api('apSignIn_day', {"linkId": "9WA12jYGulArzWS7vcrwhw", "serviceName": "dayDaySignGetRedEnvelopeSignService", "business": 1})
       if (res.data.retCode === 0) {
         console.log('签到成功')
       } else {
         console.log(res.data.retMessage)
       }
-      await wait(2000)
+      await this.wait(2000)
 
-      res = await api('signPrizeDetailList', {"linkId": "9WA12jYGulArzWS7vcrwhw", "serviceName": "dayDaySignGetRedEnvelopeSignService", "business": 1, "pageSize": 20, "page": 1})
+      res = await this.api('signPrizeDetailList', {"linkId": "9WA12jYGulArzWS7vcrwhw", "serviceName": "dayDaySignGetRedEnvelopeSignService", "business": 1, "pageSize": 20, "page": 1})
       for (let t of res.data.prizeDrawBaseVoPageBean.items) {
         if (t.prizeType === 4 && t.prizeStatus === 0) {
-          res = await api('apCashWithDraw', {"linkId": "9WA12jYGulArzWS7vcrwhw", "businessSource": "DAY_DAY_RED_PACKET_SIGN", "base": {"prizeType": t.prizeType, "business": t.business, "id": t.id, "poolBaseId": t.poolBaseId, "prizeGroupId": t.prizeGroupId, "prizeBaseId": t.prizeBaseId}})
+          res = await this.api('apCashWithDraw', {"linkId": "9WA12jYGulArzWS7vcrwhw", "businessSource": "DAY_DAY_RED_PACKET_SIGN", "base": {"prizeType": t.prizeType, "business": t.business, "id": t.id, "poolBaseId": t.poolBaseId, "prizeGroupId": t.prizeGroupId, "prizeBaseId": t.prizeBaseId}})
           console.log(parseFloat(t.prizeValue), res.data.message)
-          await wait(2000)
+          await this.wait(2000)
         }
       }
     } catch (e) {
-      console.log(e)
+      console.log('error', e)
     }
-    await wait(2000)
   }
-})()
-
-async function api(fn: string, body: object) {
-  let timestamp: number = Date.now(), t: { key: string, value: string } [] = [
-    {key: 'functionId', value: fn},
-    {key: 'body', value: SHA256(JSON.stringify(body)).toString()},
-    {key: 't', value: timestamp.toString()},
-    {key: 'appid', value: 'activities_platform'},
-    {key: 'client', value: 'H5'},
-    {key: 'clientVersion', value: '1.0.0'},
-  ]
-  let h5st = geth5st(t, '8adfb')
-  let {data} = await axios.post('https://api.m.jd.com/', `functionId=${fn}&body=${JSON.stringify(body)}&t=${timestamp}&appid=activities_platform&client=H5&clientVersion=1.0.0&h5st=${h5st}`, {
-    headers: {
-      'Host': 'api.m.jd.com',
-      'Origin': 'https://daily-redpacket.jd.com',
-      'User-Agent': 'jdltapp;',
-      'Referer': 'https://daily-redpacket.jd.com/',
-      'Content-Type': 'application/x-www-form-urlencoded',
-      'Cookie': cookie
-    }
-  })
-  return data
 }
+
+new Speed_Sign().init().then()
